requirements.txt
pytest
allure-pytest == 2.13.2
selene>=2.0.b4